% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpocket.r
\name{get_data}
\alias{get_data}
\title{Title}
\usage{
get_data(consumer_key, access_token, since = "")
}
\arguments{
\item{consumer_key}{pocket provided consumer key}

\item{access_token}{pocket provided access token}

\item{since}{unix timestamp to return items modified since date}
}
\value{
a nested list
}
\description{
Title
}
\examples{

}
